Implement 'maps:map/2' in an alternative way.
It would look like eqc_gen:function2/1 uses self() as part of its seed. This
fails in the this model as the SUT and the model gets differently executing
functions as a result. Running the map/2 in the same function and the resetting
the state by reinjection solves this problem.