Stabilize the sorting order for real() values.
The built-in sort uses '<' to sort values by default. This is a problem since we will then have
the ordering of 0 and 0.0 stable and whichever comes first in the sort order breaks everything.
Write a new sort routine, based on erts_internal:cmp_term/2 to account for this and forcefully
sort the data with this. The result is we unbreak the test case.