Improve shrinking of maps.
Rather than generating maps from vectors, generate them from resized lists. This improves
shrinking as the shrinker is now allowed to delete elements from the list and then simplify
the elements. The old solution forced a size parameter with an `eqc_gen:vector/2` but
this is much better.

Downside: we segfault 18.0 :)