Support the new badkey/badmap errors in 18.x
Split the state based on the OTP release. If we are on r18, then use a different
set of error states as the new maps explain when a bad is bad, or if the map is bad.
In places where both happen, e.g., in maps:with(0, foo), the map check takes precedence.

Passes 10 minutes of tests on both 17.5.1 and 18.0-rc1-176-gddfbbcb